+++ browser/modules/sqlQuery.js
@@ -60,8 +60,8 @@ const jquery = require('jquery');
 require('snackbarjs');
 
 let editToolsHtml = `
-        <div class="form-group gc2-edit-tools" style="display: none; width: 90%;">
-            <div class="btn-group btn-group-justified">
+        <div class="form-group gc2-edit-tools" data-edit-layer-id="{{_vidi_edit_layer_id}}" data-edit-layer-name="{{_vidi_edit_layer_name}}" data-edit-vector="{{_vidi_edit_vector}}" style="display: {{_vidi_edit_display}};">
+            <div class="btn-group btn-group-justified" style="margin: 10px 0;">
                 <div class="btn-group">
                     <button class="btn btn-primary btn-xs popup-edit-btn">
                         <i class="fa fa-pencil-alt" aria-hidden="true"></i>
@@ -82,7 +82,6 @@ let editToolsHtml = `
  */
 let defaultTemplate =
     `<div class="vidi-popup-content">
-        ${editToolsHtml}
         <h3 class="popup-title">{{_vidi_content.title}}</h3>
         {{#_vidi_content.fields}}
             {{#if value}}
@@ -239,7 +238,6 @@ module.exports = {
                 }
             }
 
-            let editingStarted = false;
             let isEmpty = true;
             let srid = metaDataKeys[value].srid;
             let key = "_vidi_sql_" + index;
@@ -263,7 +261,7 @@ module.exports = {
                 backArrowIsAdded = true;
                 defaultTemplate = `
                                 <div class='show-when-multiple-hits' style='cursor: pointer;'>
-                                    <span class='material-icons'  style=''>keyboard_arrow_left </span>
+                                    <span class='material-icons' style=''>keyboard_arrow_left </span>
                                     <span style="top: -7px;position: relative;">${__("Back")}</span>
                                 </div>` + defaultTemplate;
                 $(document).arrive('.show-when-multiple-hits', function (e, data) {
@@ -313,9 +311,11 @@ module.exports = {
 
                     isEmpty = layerObj.isEmpty();
 
-                    template = (typeof metaDataKeys[value].infowindow !== "undefined" && metaDataKeys[value].infowindow.template !== "") ? metaDataKeys[value].infowindow.template : metaDataKeys[value].type === "RASTER" ? defaultTemplateRaster : defaultTemplate;
-
-                    template = (parsedMeta.info_template && parsedMeta.info_template !== "") ? editToolsHtml + parsedMeta.info_template : template;
+                    template = metaDataKeys[value].type === "RASTER" ? defaultTemplateRaster : defaultTemplate;
+                    template = parsedMeta.info_template && parsedMeta.info_template !== "" ? parsedMeta.info_template : template;
+                    if (editingIsEnabled && layerIsEditable) {
+                        template = editToolsHtml + template;
+                    }
 
                     if (!isEmpty && !not_querable) {
 
@@ -336,10 +336,8 @@ module.exports = {
                                     .setContent(`<div id="info-box-pop-up"></div>`)
                                     .openOn(cloud.get().map)
                                     .on('remove', () => {
-                                        if (!editingStarted) {
+                                        if (!editor?.getEditedFeature()) {
                                             _self.resetAll();
-                                        } else {
-                                            editingStarted = false;
                                         }
                                     });
 
@@ -437,31 +435,6 @@ module.exports = {
                                 if (draggableEnabled) {
                                     _self.makeDraggable(popup);
                                 }
-
-                                setTimeout(() => {
-                                    if (editingIsEnabled && layerIsEditable) {
-                                        $(".gc2-edit-tools").css(`display`, `inline`);
-                                        $(".popup-edit-btn").show();
-                                        $(".popup-delete-btn").show();
-                                    } else {
-                                        $(".gc2-edit-tools").css(`display`, `none`);
-                                        $(".popup-edit-btn").hide();
-                                        $(".popup-delete-btn").hide();
-                                    }
-                                }, 100);
-
-                                $(".popup-edit-btn").unbind("click.popup-edit-btn").bind("click.popup-edit-btn", function () {
-                                    // We reset the query layer and use a unaltered layer for editor
-                                    layerObj.reset();
-                                    editor.edit(layersClone, _key_, qstore);
-                                    editingStarted = true;
-                                });
-
-                                $(".popup-delete-btn").unbind("click.popup-delete-btn").bind("click.popup-delete-btn", function () {
-                                    if (window.confirm(__(`Are you sure you want to delete the feature?`))) {
-                                        editor.delete(e, _key_, qstore);
-                                    }
-                                });
                             });
                         }
                         // Here inside onLoad we call loadDataInTable(), so the table is populated
@@ -589,6 +562,13 @@ module.exports = {
             } else {
                 fieldStr = "*";
             }
+
+            const extent = [
+                cloud.get().getExtent().left,
+                cloud.get().getExtent().bottom,
+                cloud.get().getExtent().right,
+                cloud.get().getExtent().top
+            ]
             // Get applied filters from layerTree as a WHERE clause
             let filters = layerTree.getFilterStr(keyWithoutGeom) ? layerTree.getFilterStr(keyWithoutGeom) : "1=1";
             const schemaQualifiedName = "\"" + value.split(".")[0] + "\".\"" + value.split(".")[1] + "\"";
@@ -604,18 +584,16 @@ module.exports = {
                         cloud.get().map.getSize().y,
                         cloud.get().map.latLngToContainerPoint(infoClickPoint).x,
                         cloud.get().map.latLngToContainerPoint(infoClickPoint).y,
-                        cloud.get().getExtent().left,
-                        cloud.get().getExtent().bottom,
-                        cloud.get().getExtent().right,
-                        cloud.get().getExtent().top
+                        ...extent
                     ];
                 } else {
+                    const envelope = `"${f_geometry_column}" && ST_Transform(ST_MakeEnvelope(${extent.join(',')}, 4326), ${srid})`;
                     if (geoType !== "POLYGON" && geoType !== "MULTIPOLYGON" && (!advancedInfo.getSearchOn())) {
-                        sql = "SELECT " + fieldStr + " FROM (SELECT * FROM " + schemaQualifiedName + " WHERE " + filters + ") AS foo WHERE round(ST_Distance(ST_Transform(\"" + f_geometry_column + "\"," + proj + "), ST_GeomFromText('" + wkt + "'," + proj + "))) < " + distance;
+                        sql = "SELECT " + fieldStr + " FROM (SELECT * FROM " + schemaQualifiedName + " WHERE " + filters + ") AS foo WHERE " + envelope + " AND round(ST_Distance(\"" + f_geometry_column + "\", ST_Transform(ST_GeomFromText('" + wkt + "'," + proj + ")," + srid +"))) < " + distance;
                         if (versioning) {
                             sql = sql + " AND gc2_version_end_date IS NULL ";
                         }
-                        sql = sql + " ORDER BY round(ST_Distance(ST_Transform(\"" + f_geometry_column + "\"," + proj + "), ST_GeomFromText('" + wkt + "'," + proj + ")))";
+                        sql = sql + " ORDER BY round(ST_Distance(\"" + f_geometry_column + "\", ST_Transform(ST_GeomFromText('" + wkt + "'," + proj + ")," + srid + ")))";
                     } else {
                         sql = "SELECT " + fieldStr + " FROM (SELECT * FROM " + schemaQualifiedName + " WHERE " + filters + ") AS foo WHERE ST_Intersects(ST_Transform(ST_geomfromtext('" + wkt + "'," + proj + ")," + srid + ")," + f_geometry_column + ")";
                         if (versioning) {
@@ -842,6 +820,9 @@ module.exports = {
         let metaDataKeys = meta.getMetaDataKeys();
         let parsedMeta = layerTree.parseLayerMeta(metaDataKeys[layerKey]);
         template = (parsedMeta.info_template && parsedMeta.info_template !== "") ? parsedMeta.info_template : defaultTemplate;
+        if (window.vidiConfig.enabledExtensions.includes('editor')) {
+            template = editToolsHtml + template;
+        }
         return template;
     },